backend/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0